move PermissionCodename and PERMISSIONS_CODENAMES to permConstants.ts file
